Replace `OnDomEvents.last()` with `interceptDomEvents()` action
