+++ src/dom/actions/index.ts
@@ -1,3 +1,4 @@
 export * from './capture-dom-events';
 export * from './handle-dom-events';
+export * from './intercept-dom-events';
 export * from './stop-dom-events';
